Rework VM image generation for packaging
Unfortunately cargo does not allow us to package a sub-crate in the same
crate. We also cannot use the binary artifacts feature, because that
does not allow us to set some essential Rust flags.

This work-around splits the VM image generation into a separate binary
that needs to be run at least once before publishing. Only the final VM
image will be included in the published crate.